refactor arbeidsforhold test, ekstraher test data genereing
